refactor: SwiftData model 보완
