+++ MoneyMoa/Data/SwiftDataModel/Transaction.swift
@@ -15,6 +15,7 @@ final class Transaction {
     @Attribute(.unique) var id: UUID
     var amount: Decimal
     var date: Date
+    var place: String?  // 거래 장소/대상 (맥도날드, 친구들과 더치페이, 어머니 용돈 등)
     var memo: String?
     var isFavorite: Bool
     var transactionTypeRawValue: String
@@ -30,6 +31,7 @@ final class Transaction {
         id: UUID = UUID(),
         amount: Decimal,
         date: Date = Date(),
+        place: String? = nil,
         memo: String? = nil,
         transactionType: TransactionType,
         isFavorite: Bool = false,
@@ -39,58 +41,12 @@ final class Transaction {
         self.id = id
         self.amount = amount
         self.date = date
+        self.place = place
         self.memo = memo
         self.transactionTypeRawValue = transactionType.rawValue
         self.isFavorite = isFavorite
         self.subCategory = subCategory
         self.paymentMethod = paymentMethod
-    }
-}
-
-// MARK: - Payment Method Model
-
-@Model
-final class PaymentMethod {
-    @Attribute(.unique) var id: UUID
-    var name: String
-    var kind: PaymentMethodKind
-    var orderIndex: Int
-    var isActive: Bool
-    
-    @Relationship(deleteRule: .cascade, inverse: \Transaction.paymentMethod)
-    var transactions: [Transaction]
-    
-    // Detailed information(추후 추가)
-    var institutionName: String?
-    var accountNumber: String?
-    var cardNumber: String?
-    var color: String?
-    var iconName: String?
-    
-    init(
-        id: UUID = UUID(),
-        name: String,
-        kind: PaymentMethodKind,
-        orderIndex: Int = 0,
-        isActive: Bool = true,
-        transactions: [Transaction] = [],
-        institutionName: String? = nil,
-        accountNumber: String? = nil,
-        cardNumber: String? = nil,
-        color: String? = nil,
-        iconName: String? = nil
-    ) {
-        self.id = id
-        self.name = name
-        self.kind = kind
-        self.orderIndex = orderIndex
-        self.isActive = isActive
-        self.transactions = transactions
-        self.institutionName = institutionName
-        self.accountNumber = accountNumber
-        self.cardNumber = cardNumber
-        self.color = color
-        self.iconName = iconName
     }
 }
 
@@ -103,6 +59,7 @@ extension Transaction {
             id: self.id,
             amount: self.amount,
             date: self.date,
+            place: self.place,
             memo: self.memo,
             transactionType: self.transactionType,
             isFavorite: self.isFavorite,
@@ -112,31 +69,11 @@ extension Transaction {
     }
 }
 
-extension PaymentMethod {
-    /// PaymentMethod를 PaymentMethodDTO로 변환
-    public func toDTO() -> PaymentMethodDTO {
-        return PaymentMethodDTO(
-            id: self.id,
-            name: self.name,
-            kind: self.kind,
-            orderIndex: self.orderIndex,
-            isActive: self.isActive,
-        )
-    }
-}
-
 // MARK: - Collection Extensions
 
 extension Collection where Element == Transaction {
     /// Transaction 배열을 TransactionDTO 배열로 변환
     func toDTOs() -> [TransactionDTO] {
-        return self.map { $0.toDTO() }.sorted()
-    }
-}
-
-extension Collection where Element == PaymentMethod {
-    /// PaymentMethod 배열을 PaymentMethodDTO 배열로 변환
-    func toDTOs() -> [PaymentMethodDTO] {
         return self.map { $0.toDTO() }.sorted()
     }
 }
@@ -153,24 +90,12 @@ extension TransactionDTO {
             id: self.id,
             amount: self.amount,
             date: self.date,
+            place: self.place,
             memo: self.memo,
             transactionType: self.transactionType,
             isFavorite: self.isFavorite,
             subCategory: subCategory,
             paymentMethod: paymentMethod
-        )
-    }
-}
-
-extension PaymentMethodDTO {
-    /// PaymentMethodDTO를 SwiftData PaymentMethod 모델로 변환
-    func toModel() -> PaymentMethod {
-        return PaymentMethod(
-            id: self.id,
-            name: self.name,
-            kind: self.kind,
-            orderIndex: self.orderIndex,
-            isActive: self.isActive,
         )
     }
 }
